+++ apps/ledger-live-desktop/tests/specs/services/ethereumStaking.spec.ts
@@ -115,7 +115,7 @@ test("Ethereum staking flows via portfolio, asset page and market page @smoke", 
   });
 
   await test.step("choose ethereum account", async () => {
-    await drawer.selectAccount("Ethereum", 1);
+    await drawer.selectAccount("Ethereum", 0);
     await expect.soft(page).toHaveScreenshot("choose-stake-provider-modal-from-portfolio-page.png");
   });
 
@@ -124,7 +124,7 @@ test("Ethereum staking flows via portfolio, asset page and market page @smoke", 
       event: "button_clicked2",
       properties: {
         button: "kiln",
-        path: "account/mock:1:ethereum:true_ethereum_1:",
+        path: "account/mock:1:ethereum:true_ethereum_0:",
         modal: "stake",
         flow: "stake",
         value: "/platform/kiln",
@@ -132,7 +132,7 @@ test("Ethereum staking flows via portfolio, asset page and market page @smoke", 
     });
     await delegate.chooseStakeProvider("kiln");
     await analyticsPromise;
-    await liveAppWebview.waitForCorrectTextInWebview("Ethereum 2");
+    await liveAppWebview.waitForCorrectTextInWebview("Ethereum 1");
     const dappURL = await liveAppWebview.getLiveAppDappURL();
     expect(await liveAppWebview.getLiveAppTitle()).toBe("Kiln");
     expect(dappURL).toContain("?focus=dedicated");
@@ -151,7 +151,7 @@ test("Ethereum staking flows via portfolio, asset page and market page @smoke", 
     await assetPage.startStakeFlow();
     await drawer.waitForDrawerToBeVisible();
     await expect.soft(page).toHaveScreenshot("stake-drawer-opened-from-asset-page.png");
-    await drawer.close();
+    await drawer.closeDrawer();
   });
 
   await test.step("start stake flow via Account page", async () => {
@@ -181,7 +181,7 @@ test("Ethereum staking flows via portfolio, asset page and market page @smoke", 
     await expect
       .soft(page)
       .toHaveScreenshot("stake-drawer-opened-from-market-page.png", maskPartOfItemsInMarket);
-    await drawer.close();
+    await drawer.closeDrawer();
   });
 
   await test.step("Go back to Market page and start stake from ETH coin detail page", async () => {
@@ -191,12 +191,12 @@ test("Ethereum staking flows via portfolio, asset page and market page @smoke", 
     await marketCoinPage.startStakeFlow();
     await drawer.waitForDrawerToBeVisible();
     await expect.soft(page).toHaveScreenshot("stake-drawer-opened-from-market-coin-page.png");
-    await drawer.selectAccount("Ethereum", 1);
+    await drawer.selectAccount("Ethereum", 0);
     const analyticsPromise = analytics.waitForTracking({
       event: "button_clicked2",
       properties: {
         button: "kiln_pooling",
-        path: "account/mock:1:ethereum:true_ethereum_0:",
+        path: "account/mock:1:ethereum:true_ethereum_1:",
         modal: "stake",
         flow: "stake",
         value: "/platform/kiln",
@@ -205,7 +205,7 @@ test("Ethereum staking flows via portfolio, asset page and market page @smoke", 
     await delegate.chooseStakeProvider("kiln_pooling");
     await analyticsPromise;
     const dappURL = await liveAppWebview.getLiveAppDappURL();
-    await liveAppWebview.waitForCorrectTextInWebview("Ethereum 1");
+    await liveAppWebview.waitForCorrectTextInWebview("Ethereum 2");
     expect(dappURL).toContain("?focus=pooled");
     expect(await liveAppWebview.getLiveAppTitle()).toBe("Kiln");
   });