+++ apps/ledger-live-mobile/src/components/RootNavigator/BaseNavigator.tsx
@@ -209,7 +209,6 @@ export default function BaseNavigator() {
           options={{
             headerStyle: styles.headerNoShadow,
           }}
-          {...noNanoBuyNanoWallScreenOptions}
         />
         <Stack.Screen
           name={ScreenName.Recover}