+++ apps/ledger-live-mobile/src/const/navigation.ts
@@ -541,6 +541,7 @@ export enum ScreenName {
   SelectDevice = "SelectDevice",
   SelectAccounts = "SelectAccounts",
   ScanDeviceAccounts = "ScanDeviceAccounts",
+  AddAccountsWarning = "AddAccountsWarning",
 }
 
 export enum NavigatorName {