+++ apps/ledger-live-mobile/src/newArch/features/Accounts/Navigator.tsx
@@ -2,22 +2,25 @@ import React, { useCallback, useMemo } from "react";
 import { Platform } from "react-native";
 import { createStackNavigator } from "@react-navigation/stack";
 import { useTheme } from "styled-components/native";
-import { useRoute } from "@react-navigation/native";
+import { useNavigation, useRoute } from "@react-navigation/native";
 import { ScreenName } from "~/const";
 import { getStackNavigatorConfig } from "~/navigation/navigatorConfig";
 import { track } from "~/analytics";
 import type { NetworkBasedAddAccountNavigator } from "LLM/features/Accounts/screens/AddAccount/types";
-import SelectAccounts from "LLM/features/Accounts/screens/SelectAccounts";
 import { NavigationHeaderCloseButtonAdvanced } from "~/components/NavigationHeaderCloseButton";
 import ScanDeviceAccounts from "LLM/features/Accounts/screens/ScanDeviceAccounts";
 import { AccountsListNavigator } from "./screens/AccountsList/types";
 import { useFeature } from "@ledgerhq/live-common/featureFlags/index";
 import AccountsList from "LLM/features/Accounts/screens/AccountsList";
+import { NavigationHeaderBackButton } from "~/components/NavigationHeaderBackButton";
+import AddAccountsSuccess from "./screens/AddAccountSuccess";
+import SelectAccounts from "./screens/SelectAccounts";
 
 export default function Navigator() {
   const { colors } = useTheme();
   const route = useRoute();
   const accountListUIFF = useFeature("llmAccountListUI");
+  const navigation = useNavigation();
 
   const onClose = useCallback(() => {
     track("button_clicked", {
@@ -34,6 +37,14 @@ export default function Navigator() {
     [colors, onClose],
   );
 
+  const onPressBack = useCallback(() => {
+    track("button_clicked", {
+      button: "Back",
+      page: route.name,
+    });
+    navigation.goBack();
+  }, [route, navigation]);
+
   return (
     <Stack.Navigator
       screenOptions={{
@@ -41,6 +52,14 @@ export default function Navigator() {
         gestureEnabled: Platform.OS === "ios",
       }}
     >
+      {/* Scan accounts from device */}
+      <Stack.Screen
+        name={ScreenName.ScanDeviceAccounts}
+        component={ScanDeviceAccounts}
+        options={{
+          headerTitle: "",
+        }}
+      />
       {/* Select Accounts */}
       <Stack.Screen
         name={ScreenName.SelectAccounts}
@@ -50,24 +69,24 @@ export default function Navigator() {
         }}
         initialParams={route.params}
       />
-
-      {/* Scan accounts from device */}
-      <Stack.Screen
-        name={ScreenName.ScanDeviceAccounts}
-        component={ScanDeviceAccounts}
-        options={{
-          headerTitle: "",
-        }}
-      />
       {accountListUIFF?.enabled && (
         <Stack.Screen
           name={ScreenName.AccountsList}
           component={AccountsList}
           options={{
             headerTitle: "",
+            headerLeft: () => <NavigationHeaderBackButton onPress={onPressBack} />,
           }}
         />
       )}
+      <Stack.Screen
+        name={ScreenName.AddAccountsSuccess}
+        component={AddAccountsSuccess}
+        options={{
+          headerTitle: "",
+          headerLeft: () => null,
+        }}
+      />
     </Stack.Navigator>
   );
 }