+++ apps/ledger-live-mobile/src/newArch/features/DeviceSelection/Navigator.tsx
@@ -3,16 +3,13 @@ import { Platform } from "react-native";
 import { createStackNavigator } from "@react-navigation/stack";
 import { useTheme } from "styled-components/native";
 import { useTranslation } from "react-i18next";
-import { NavigationProp, useRoute } from "@react-navigation/native";
+import { useRoute } from "@react-navigation/native";
 import { ScreenName } from "~/const";
 import { getStackNavigatorConfig } from "~/navigation/navigatorConfig";
 import { track } from "~/analytics";
 import SelectDevice, {
   addAccountsSelectDeviceHeaderOptions,
 } from "LLM/features/DeviceSelection/screens/SelectDevice";
-import ConnectDevice, {
-  connectDeviceHeaderOptions,
-} from "LLM/features/DeviceSelection/screens/ConnectDevice";
 import StepHeader from "~/components/StepHeader";
 import { NavigationHeaderCloseButtonAdvanced } from "~/components/NavigationHeaderCloseButton";
 import { DeviceSelectionNavigatorParamsList } from "./types";
@@ -36,14 +33,6 @@ export default function Navigator() {
     }),
     [colors, onClose],
   );
-
-  const onConnectDeviceBack = useCallback((navigation: NavigationProp<Record<string, unknown>>) => {
-    track("button_clicked", {
-      button: "Back arrow",
-      page: ScreenName.ConnectDevice,
-    });
-    navigation.goBack();
-  }, []);
 
   return (
     <Stack.Navigator
@@ -71,23 +60,7 @@ export default function Navigator() {
         initialParams={route.params}
       />
 
-      {/* Select / Connect Device */}
-      <Stack.Screen
-        name={ScreenName.ConnectDevice}
-        component={ConnectDevice}
-        options={({ navigation }) => ({
-          headerTitle: () => (
-            <StepHeader
-              subtitle={t("transfer.receive.stepperHeader.range", {
-                currentStep: "2",
-                totalSteps: 3,
-              })}
-              title={t("transfer.receive.stepperHeader.connectDevice")}
-            />
-          ),
-          ...connectDeviceHeaderOptions(() => onConnectDeviceBack(navigation)),
-        })}
-      />
+      {/*Connect Device : Only for receive flow context it will be re-added & adjusted in https://ledgerhq.atlassian.net/browse/LIVE-14726 */}
     </Stack.Navigator>
   );
 }