+++ apps/ledger-live-mobile/src/screens/AddAccounts/01-SelectCrypto.tsx
@@ -57,6 +57,8 @@ export default function AddAccountsSelectCrypto({ navigation, route }: Props) {
 
   const mock = useEnv("MOCK");
 
+  const aptos = useFeature("currencyAptos");
+  const aptosTestnet = useFeature("currencyAptosTestnet");
   const axelar = useFeature("currencyAxelar");
   const stargaze = useFeature("currencyStargaze");
   const secretNetwork = useFeature("currencySecretNetwork");
@@ -108,9 +110,12 @@ export default function AddAccountsSelectCrypto({ navigation, route }: Props) {
   const zksyncSepolia = useFeature("currencyZkSyncSepolia");
   const mantra = useFeature("currencyMantra");
   const xion = useFeature("currencyXion");
+  const zenrock = useFeature("currencyZenrock");
 
   const featureFlaggedCurrencies = useMemo(
     (): Partial<Record<CryptoCurrencyId, Feature<unknown> | null>> => ({
+      aptos,
+      aptos_testnet: aptosTestnet,
       axelar,
       stargaze,
       umee,
@@ -162,8 +167,11 @@ export default function AddAccountsSelectCrypto({ navigation, route }: Props) {
       zksync_sepolia: zksyncSepolia,
       mantra,
       xion,
+      zenrock,
     }),
     [
+      aptos,
+      aptosTestnet,
       axelar,
       stargaze,
       umee,
@@ -215,6 +223,7 @@ export default function AddAccountsSelectCrypto({ navigation, route }: Props) {
       zksyncSepolia,
       mantra,
       xion,
+      zenrock,
     ],
   );