+++ libs/coin-modules/coin-xrp/src/logic/listOperations.test.ts
@@ -1,9 +1,12 @@
+import { assert } from "console";
 import { listOperations } from "./listOperations";
 import { RIPPLE_EPOCH } from "./utils";
 
+const maxHeight = 2;
+const minHeight = 1;
 const mockGetServerInfos = jest.fn().mockResolvedValue({
   info: {
-    complete_ledgers: "1-2",
+    complete_ledgers: `${minHeight}-${maxHeight}`,
   },
 });
 const mockGetTransactions = jest.fn();
@@ -16,6 +19,94 @@ describe("listOperations", () => {
   afterEach(() => {
     mockGetServerInfos.mockClear();
     mockGetTransactions.mockClear();
+  });
+
+  it("when there are no transactions then the result is empty", async () => {
+    // Given
+    mockGetTransactions.mockResolvedValue([]);
+    // When
+    const [results, token] = await listOperations("any address", { minHeight: 0 });
+    // Then
+    expect(mockGetServerInfos).toHaveBeenCalledTimes(1);
+    expect(mockGetTransactions).toHaveBeenCalledTimes(1);
+    expect(results).toEqual([]);
+    expect(token).toEqual(minHeight);
+  });
+
+  it("when there are no transactions and a limit then the result is empty", async () => {
+    // Given
+    mockGetTransactions.mockResolvedValue([]);
+    // When
+    const [results, token] = await listOperations("any address", { minHeight: 0, limit: 1 });
+    // Then
+    expect(mockGetServerInfos).toHaveBeenCalledTimes(1);
+    expect(mockGetTransactions).toHaveBeenCalledTimes(1);
+    expect(results).toEqual([]);
+    expect(token).toEqual(minHeight);
+  });
+
+  const paymentTx = {
+    ledger_hash: "HASH_VALUE_BLOCK",
+    hash: "HASH_VALUE",
+    close_time_iso: "2000-01-01T00:00:01Z",
+    meta: { delivered_amount: "100" },
+    tx_json: {
+      TransactionType: "Payment",
+      Fee: "1",
+      ledger_index: 2,
+      date: 1000,
+      Account: "sender",
+      Destination: "dest",
+      Sequence: 1,
+    },
+  };
+  const otherTx = { ...paymentTx, tx_json: { ...paymentTx.tx_json, TransactionType: "Other" } };
+
+  it("should only list operations of type payment", async () => {
+    // Given
+    const lastTransaction = paymentTx;
+    mockGetTransactions.mockResolvedValueOnce([paymentTx, otherTx, lastTransaction]);
+    mockGetTransactions.mockResolvedValue([]); // subsequent calls
+
+    // When
+    const [results, token] = await listOperations("any address", { minHeight: 0, limit: 3 });
+
+    // Then
+    expect(mockGetServerInfos).toHaveBeenCalledTimes(1);
+    // it's called twice because first call yields only 2 transactions, and 3 are asked
+    expect(mockGetTransactions).toHaveBeenCalledTimes(2);
+    expect(results.length).toEqual(2);
+    expect(token).toEqual(lastTransaction.tx_json.ledger_index - 1);
+  });
+
+  it("should make enough calls so that the limit requested is satified", async () => {
+    const lastTransaction = otherTx;
+    const txs = [paymentTx, paymentTx, otherTx, otherTx, otherTx, otherTx, otherTx, otherTx];
+    assert(txs.length === 8);
+    mockGetTransactions.mockResolvedValue(txs);
+
+    const [results, token] = await listOperations("any address", { minHeight: 0, limit: 8 });
+
+    expect(mockGetServerInfos).toHaveBeenCalledTimes(1);
+    // it's called 4 times because each call yields only 2 transactions, and 8 are asked
+    expect(mockGetTransactions).toHaveBeenCalledTimes(4);
+    expect(results.length).toEqual(8);
+    expect(token).toEqual(lastTransaction.tx_json.ledger_index - 1);
+  });
+
+  it("should make enough calls, even if there is not enough txs to satisfy the limit", async () => {
+    mockGetTransactions.mockResolvedValueOnce([otherTx, otherTx, otherTx, otherTx]);
+    mockGetTransactions.mockResolvedValueOnce([paymentTx, paymentTx]);
+    mockGetTransactions.mockResolvedValue([]); // subsequent calls
+    const lastTransaction = paymentTx;
+
+    const [results, token] = await listOperations("any address", { minHeight: 0, limit: 4 });
+
+    expect(mockGetServerInfos).toHaveBeenCalledTimes(1);
+    // it's called 2 times because the second call is a shortage of txs
+    expect(mockGetTransactions).toHaveBeenCalledTimes(2);
+    expect(results.length).toEqual(2);
+    expect(token).toEqual(lastTransaction.tx_json.ledger_index - 1);
   });
 
   it.each([
@@ -34,7 +125,7 @@ describe("listOperations", () => {
   ])(
     "should return the list of operations associated to an account",
     async ({ address, opSender, opDestination, expectedType }) => {
-      // Givem
+      // Given
       const deliveredAmount = 100;
       const fee = 10;
       mockGetTransactions.mockResolvedValue([
@@ -95,7 +186,7 @@ describe("listOperations", () => {
       ]);
 
       // When
-      const [results, _] = await listOperations(address, { startAt: 0 });
+      const [results, _] = await listOperations(address, { minHeight: 0 });
 
       // Then
       expect(mockGetServerInfos).toHaveBeenCalledTimes(1);