+++ libs/ledger-live-common/src/e2e/families/aptos.ts
@@ -1,0 +1,7 @@
+import { pressBoth, pressUntilTextFound } from "../speculos";
+import { DeviceLabels } from "../enum/DeviceLabels";
+
+export async function sendAptos() {
+  await pressUntilTextFound(DeviceLabels.APPROVE);
+  await pressBoth();
+}